Replace 'IDE Key' terminology with 'Debug Trigger' (#7)
* Use 'debug trigger' instead of 'IDE Key'

The 'IDE Key' concept hasn't been a thing for a long time, and perpetuating it
through hints and documentation is not a good idea. The pre-defined values also
make no sense, as they should refer to a **user's** name, and not an IDE name.

It is now called 'trigger' and the same cooking value can be used for Xdebug
Cloud.

* Remove trailing whitespace in .md file

* Continue renaming the IDE Key to Debug Trigger

* Show default value in placeholder